digits of profits do not match, huge problem with profitGiven
We have to find better way of keeping record of profitGiven because we need to calculate it for each investment made to the same projectID. I have some ideas but it includes deletion of projects. Also there is a big amount of digits differences between first half and second half calculation. Not because of denominator. Tests will be organized after the changes are made.